libimage: disk usage: catch corrupted images
Make sure to check an image for corruption before running disk usage on
it.  Such checks are already applied on various execution paths but not
yet on disk usage.

Further update the corrupted-image error to include that the image
should be removed to resolve the error.  This should ultimately guide
users to resolve the issue.

Fixes: containers/common#751
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>