Backport store: lock graphLock during to 1.36
Backport the fix for #1037 to
the release-1.36 branch so that it can then be backported to Podman's
v3.4.2-rhel branch for RHEL 8.5.0.2

Note: containers/storage#1097 was completed to the 1.37 branch in error. Podman
v3.4.2-rhel is still on c/storage v1.36.  #1098 was submited to main
in error, hopefully the third time is the charm?

Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>